labs/Lab8CodeSolution/resumebuilder/resumeapp/serializers.py
from rest_framework import serializers
from .models import Qualification, WorkExperience

class QualificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Qualification
        fields = '__all__'

class WorkExperienceSerializer(serializers.ModelSerializer):
    class Meta:
        model = WorkExperience
        fields = '__all__'